moving tests out from folders
